orders-dashboard/requirement.txt
# Core Dependencies
pandas>=1.3.0
streamlit>=1.20.0
altair>=5.0.0
mysql-connector-python>=8.0.33

# Optional but recommended
python-dotenv>=0.21.0

# Testing (optional)
pytest>=7.0
pytest-cov>=4.0

# Deprecated but kept for backward compatibility
# fpdf (not used in current version)
# matplotlib (replaced by altair)